GUACAMOLE-1218: Refactor guacamole-auth-json's StringListProperty into guacamole-ext.
