+++ extensions/guacamole-auth-json/src/main/java/org/apache/guacamole/auth/json/ConfigurationService.java
@@ -25,6 +25,7 @@ import java.util.Collections;
 import org.apache.guacamole.GuacamoleException;
 import org.apache.guacamole.environment.Environment;
 import org.apache.guacamole.properties.ByteArrayProperty;
+import org.apache.guacamole.properties.StringListProperty;
 
 /**
  * Service for retrieving configuration information regarding the JSON
